Task: 005 = Move models to Api.Contracts and update endpoint
