+++ TimeWarp.Architecture/Source/ContainerApps/Api/Api.Server/Features/WeatherForecast/WeatherEndpoint.cs
@@ -6,7 +6,7 @@ namespace TimeWarp.Architecture.Features.WeatherForecasts;
 /// <remarks>
 /// Gets Weather Forecasts for the number of days specified in the request
 /// </remarks>
-public class WeatherEndpoint : EndpointWithoutRequest<IEnumerable<WeatherResponse>>
+public class WeatherEndpoint : Endpoint<WeatherRequest, IEnumerable<WeatherResponse>>
 {
   private static readonly string[] Summaries =
   [
@@ -37,10 +37,11 @@ public class WeatherEndpoint : EndpointWithoutRequest<IEnumerable<WeatherRespons
 
   public override async Task HandleAsync
   (
+    WeatherRequest aRequest,
     CancellationToken aCancellationToken
   )
   {
-    int days = Query<int?>("days") ?? 5;
+    int days = aRequest.Days ?? 5;
 
     if (days <= 0)
     {
@@ -60,47 +61,5 @@ public class WeatherEndpoint : EndpointWithoutRequest<IEnumerable<WeatherRespons
     );
 
     await SendAsync(forecasts, cancellation: aCancellationToken);
-  }
-}
-
-/// <summary>
-/// Weather forecast response
-/// </summary>
-public sealed class WeatherResponse
-{
-  /// <summary>
-  /// The forecast for this Date
-  /// </summary>
-  /// <example>2020-06-08T12:32:39.9828696+07:00</example>
-  public DateTime Date { get; init; }
-
-  /// <summary>
-  /// Summary of the forecast
-  /// </summary>
-  /// <example>Cool</example>
-  public string Summary { get; init; }
-
-  /// <summary>
-  /// Temperature in Celsius
-  /// </summary>
-  /// <example>24</example>
-  public int TemperatureC { get; init; }
-
-  /// <summary>
-  /// Temperature in Fahrenheit
-  /// </summary>
-  /// <example>75</example>
-  public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-
-  public WeatherResponse
-  (
-    DateTime date,
-    string summary,
-    int temperatureC
-  )
-  {
-    Date = date;
-    Summary = summary;
-    TemperatureC = temperatureC;
   }
 }
